Minor updates to several python scripts
