+++ python/platosim/photometry.py
@@ -13,18 +13,11 @@ import glob
 import h5py
 import scipy
 import pathlib
-
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
-
 from scipy import constants as c
-from zipfile import ZipFile
-from tqdm import tqdm
-#from numba import njit
-#from numba_progress import ProgressBar
 
-import platosim.plot            as pt
 import platosim.utilities       as ut
 import platosim.referenceFrames as rf
 
@@ -93,467 +86,9 @@ class Photometry(object):
         return
 
 
-    #--------------------------------------------------------------#
-    #                          PREPARE DATA                        #
-    #--------------------------------------------------------------#
-
-
-    def files(self, extension="zip", path=None):
-        """
-        PURPOSE: Function to unpack all files multi-camera and multi-quarter
-                 light curves simulated for a single star.
-        """
-
-        # Set path to files as default
-        if path is None: path = self.path
-
-        # Fetch all zip files
-        files = glob.glob(f"{path}/*.{extension}")
-
-        return files
-
-    
-    
-
-    def unpack(self, files=None):
-        """
-        PURPOSE: Function to unpack all files multi-camera and multi-quarter
-                 light curves simulated for a single star.
-        """
-
-        # Fetch all zip files
-        if files is None:
-            files = glob.glob(self.path + "/*.zip")
-
-        # Unpack zip files
-        for f in files:
-            with ZipFile(f, 'r') as unzip:
-                unzip.extractall(self.path)
-
-
     
     #--------------------------------------------------------------#
     #                      SINGLE CAMERA/QUARTER                   #
     #--------------------------------------------------------------#
         
 
-    def getTime(self):
-
-        """
-        Returns time points.
-        """
-
-        # Check if column exist in HSD5 file
-        if "Time" not in self.hdf5file["StarPositions"].keys():
-            ut.errorcode("warning","getTime(): Time array not present in StarPositions group!")
-            return None
-        else:
-            # Fetch time column and write to data frame
-            time = np.array(self.hdf5file["StarPositions/Time"])
-            return pd.DataFrame({"time": time})
-
-
-
-    
-    def getFlux(self, starID, flux_type="estimated"):
-
-        """Returns flux points.
-        """
-
-        # Select the proper flux name
-
-        if   flux_type == "estimated": lctype = "estimatedFlux"
-        elif flux_type == "input":     lctype = "inputFlux"
-        else:
-            ut.errorcode("error", "getFlux(): flux_type can only be 'estimated' or 'input'")
-
-        # Query either a single star or multiple stars as requested
-
-        try: len(starID)
-        except:
-            starID = np.array([starID])
-            names = False
-        else:
-            names = True
-            
-        # Add a flux column(s) for the star(s)
-        
-        for ID in starID:
-            
-            # Check photometry is present for each star
-            
-            starIDgroupName = f"starID{ID}"
-            if starIDgroupName not in self.hdf5file["Photometry"]["Lightcurves"].keys():
-                ut.errorcode("error", f"getLightCurve(): {starIDgroupName} not present in" +
-                                       "Photometry/Lightcurves/ in the HDF5 file")
-
-            # Select correct name convention
-
-            if names: string = f"flux_{ID}"
-            else:     string = "flux"
-                
-            # Fetch flux column
-            
-            flux = np.array(self.hdf5file[f"Photometry/Lightcurves/starID{ID}/{lctype}"])
-
-            # Create data frame and append to it
-
-            if ID == starID[0]:
-                df = pd.DataFrame({string: flux})
-            else:
-                df[string] = flux
-            
-        # Finito!
-            
-        return df
-
-
-    
-    
-
-    
-    def getLightCurve(self, starID, flux_type="estimated"):
-
-        """Extract the light curve of one or more stars
-
-        This function can be used to extract the light curves for one star
-        with the given ID, or for a full list of stars with IDs given by an
-        list or array.
-
-        Parameters
-        ----------
-        starID : int, list/ndarray
-            int  : ID of the star as mentioned in the last column of the star catalog file
-            list : List of star IDs for which the light curve should be extracted
-        flux_type : str
-            Either "estimated" or "input". 
-            The estimated one is derived from a binary mask.
-            The input one is derived from the mean input magnitude specified in the star catalog
-            and (for variable stars) the delta-magnitude time series given as an input file.
-
-        Return
-        ------
-        df : pandas data frame
-            time : first column [s] 
-            flux : consecutive columns [e-/exposure]
-        """
-
-        # Fetch time column
-        
-        time = self.getTime()
-        
-        # Fetch flux column(s)
-
-        flux = self.getFlux(starID, flux_type=flux_type)
-
-        # Combine data
-
-        return pd.concat([time, flux], axis=1)
-
-
-
-
-
-    def getMaskUpdateEvents(self):
-
-        """Exposure number of all mask updates.
-        """
-        
-        # Fetch mask update events
-
-        return np.array(self.hdf5file["Photometry/Masks/exposureNrOfMaskUpdate"])
-
-
-
-    
-    
-
-    def getApertureMask(self, starID, imageNr=None):
-
-        """Fetch all information about the photometric aperture mask used.
-
-        This function returns the subfield row and column indices of the mask that
-        is used to extract the flux of star with the given ID for the given exposure
-        number. This only makes sense if the photometry was activated in the configuration
-        yaml file.
-
-        INPUT: 
-        starID:  ID of the star as mentioned in the last column of the star catalog file
-               imageNr: integer sequential number of the image in the HDF5 file
-
-        OUTPUT: rowIndices: subimage row indices for each of the mask pixels
-                colIndices: subimage column indices for each of the mask pixels
-                exposureNr: the image number in which the mask was derived.
-                            exposureNr <= imageNr
-
-        NOTE: Masks are not update continuously, but only once in a while. This function searches
-              for the most recent mask. This mask may have thus been derived from a previous image
-              rather than from the given image Nr.
-        """
-
-
-        # Check if photometric data exists
-
-        starIDgroupName = "starID{0}".format(starID)
-        if starIDgroupName not in self.hdf5file["Photometry"]["Masks"].keys():
-            print("Error: getPhotometricMask(): " + starIDgroupName + " not present in Photometry/Masks/ in the HDF5 file")
-            return None, None, None, None, None
-
-        # Fetch mask info and mask updates
-
-        mask = self.hdf5file["Photometry"]["Masks"]
-        exposureNrOfMaskUpdate = np.array(mask["exposureNrOfMaskUpdate"])
-        numMaskUpdates = len(exposureNrOfMaskUpdate)
-
-        # If a specific image number for the mask update is requested:
-        # NOTE Masks are not updated for every exposure hence find most recent mask
-
-        if isinstance(imageNr, int):
-
-            idx = np.searchsorted(exposureNrOfMaskUpdate, imageNr, side='right') - 1
-            if idx < 0:
-                print("Error: getPhotometricMask(): requesting an imageNr that is too early for this HDF5 file")
-                return None, None, None, None, None
-
-            exposureNrOfMaskUpdate = exposureNrOfMaskUpdate[idx]
-
-            # Extact mask size and NSR
-
-            maskSize = np.array(mask[starIDgroupName]['maskSize'])[idx]
-            maskNSR  = np.array(mask[starIDgroupName]['maskNSR'])[idx]
-
-            # Extract the indices of the proper mask
-
-            exposureGroupName = "Exposure{0:06d}".format(exposureNrOfMaskUpdate)
-            rowIndices = np.array(mask[starIDgroupName][exposureGroupName]["maskRowIndices"])
-            colIndices = np.array(mask[starIDgroupName][exposureGroupName]["maskColumnIndices"])
-
-        # Else fetch all the indices for all mask updates
-
-        else:
-
-            # Extact mask size and NSR
-
-            maskSize = np.array(mask[starIDgroupName]['maskSize'])
-            maskNSR  = np.array(mask[starIDgroupName]['maskNSR'])
-
-            # Extract the indices of all masks
-
-            rowIndices = []
-            colIndices = []
-            for i in range(numMaskUpdates):
-                exposureGroupName = "Exposure{0:06d}".format(exposureNrOfMaskUpdate[i])
-                rowIndices.append(mask[starIDgroupName][exposureGroupName]["maskRowIndices"])
-                colIndices.append(mask[starIDgroupName][exposureGroupName]["maskColumnIndices"])
-            rowIndices = np.array(rowIndices)
-            colIndices = np.array(colIndices)
-
-        # Finito!
-
-        return rowIndices, colIndices, exposureNrOfMaskUpdate, maskSize, maskNSR
-
-
-    
-
-    
-
-
-
-    def plotLightCurve(self, starID, medfilt=144, figsize=(8,5)):
-
-        """Plot the photometric time series of single target from a HDF5 file.
-
-        Parameters
-        ----------
-        outputFile : str
-            File name of HDF5 file containing photometry
-        medfilt : int
-            Number of time points for overlaid median filter.
-            Default is 1h cadence: 3600s/25s=144 time points.
-        NSR : bool
-            If the Noise-to-Signal Ratio (NSR) should be plotted alonside the time series.
-            Default is False.
-        COB : bool
-            If the Center-Of-Brightness (COB) should be plotted alonside the time series.
-            Default is False.
-        title : str
-            String of title plot. Default is False.
-
-        Return
-        ------
-        fig, ax : objects
-            Axes matplotlib.pyplot handle object to be modified by the user
-        """
-
-        # Fetch photometry
-
-        df = self.getLightCurve(starID)
-        df["time"] /= 86400   # [days]
-
-        # Compute median carbox filter of output signal
-
-        # if flux_med:
-        #     flux_med = median_filter(flux_out, medfilt)  # [ppm]
-
-        # Fetch mask updates
-
-        #mask = f.getPhotometricMask(1)
-        #maskupdates = (mask[2] * 25.) / c.day   # [days]
-        #maskNSR     = mask[4] * 1e6             # [ppm]
-
-        # Create figure and subplots
-
-        return pt.plotPhotometry(df, time_unit="days", flux_unit="e/s")
-
-        
-
-        #plt.subplots_adjust(wspace=0.15, hspace=0.20)
-
-
-        # #axes_maskupdates(ax0, time, maskupdates)
-        # ax.legend(loc=legendLocation) #, fancybox=True, ncol=2)
-        # ax.set_ylabel('Flux [e$^-$ s$^{-1}$]')
-
-        # # Plot NSR
-
-        # if NSR is not False:
-
-        #     ax1 = fig.add_subplot(2,1,2)
-        #     ax1.plot(maskupdates, maskNSR, 'k--', alpha=0.5)
-        #     ax1.plot(maskupdates, maskNSR, 'm*', label='Mask update NSR')
-        #     ax1.legend(loc='upper right',  fancybox=True)
-        #     ax1.set_xlabel('Time [days]')
-        #     ax1.set_ylabel(r'NSR [ppm h$^{-1}$]')
-        #     ax1.set_xlim(axes_minmax(x=time))
-
-        # if COB is not False:
-
-        #     # Fetch pixel coordinates
-
-        #     rowPix, colPix = f.getStarPositions(1)
-
-        #     # Plot row pixel on left y axis
-
-        #     ax1 = fig.add_subplot(2,1,2)
-        #     ax1.plot(time, rowPix, '-', c='darkcyan', label='Row pixel')
-        #     ax1.set_ylabel('Row coordinate [pixel]')
-        #     ax1.set_xlabel('Time [days]')
-        #     ax1.set_title('Star position')
-        #     ax1.set_xlim(axes_minmax(x=time))
-        #     ax1.legend(loc='upper left')
-
-        #     # Plot column pixel on right y axis
-
-        #     ax2 = ax1.twinx()
-        #     ax2.plot(time, colPix, '-', c='hotpink', label='Col pixel')
-        #     ax2.set_ylabel('Column coordinate [pixel]')
-        #     ax2.set_xlim(axes_minmax(x=time))
-        #     ax2.legend(loc='lower right')
-
-        # # Labels
-
-        # if title is not False:
-        #     fig.text(0.5, 0.9, title, ha='center', fontsize=fs)
-        # if NSR is False and COB is False:
-        #     ax.set_xlabel('Time [days]')
-
-        # # Limits
-
-        # #flux_min = np.min([flux_in, flux_out])
-        # #ax.set_ylim(flux_min, np.abs(flux_min))
-        # #ax.set_xlim(axes_minmax(x=time))
-        # plt.show()
-        
-        # # Finito!
-
-        # return fig, ax
-
-        
-
-
-
-    
-
-    def getPhotometry(self, starID, quarterNo=1):
-        """
-        PURPOSE: Extract the light curve of the star with the given ID.
-        NOTE The estimated one is derived from a binary mask. The input one is derived
-        from the mean input magnitude specified in the star catalog and
-        (for variable stars) the delta-mag time series given as an input file.
-
-        PARAMETERS
-        ----------
-        starID : int
-            ID of the star as mentioned in the last column of the star catalog file
-        quarterNo : int
-            Mission quarter number used to return correct relative time points.
-
-        RETURN
-        ------
-        time : ndarray
-            Time points of time series [s]
-        flux : ndarray
-            Flux points of time series [electrons/exposure]
-        """
-
-        # Verify if the necessary info is in the HDF5 file
-
-        if "Photometry" not in self.hdf5file["/"].keys():
-            print("Error: getLightCurve(): No photometry present in the HDF5 file")
-            return None, None
-
-        starIDgroupName = "starID{0}".format(starID)
-        if starIDgroupName not in self.hdf5file["Photometry"]["Lightcurves"].keys():
-            print("Error: getLightCurve(): " + starIDgroupName + " not present in Photometry/Lightcurves/ in the HDF5 file")
-            return None, None
-
-        # Fetch time points
-
-        time = self.getTimeQuarter(quarterNo)
-
-        # Extract the flux and the time points
-
-        photometry = self.hdf5file["Photometry"]["Lightcurves"][starIDgroupName]
-        inputFlux     = np.array(photometry["inputFlux"])
-        estimatedFlux = np.array(photometry["estimatedFlux"])
-
-        return time, inputFlux, estimatedFlux
-
-
-
-    
-
-    def getTimeQuarter(self, quarterNo):
-
-        """Function to create and return quarter-long time column.
-
-        This function creates a time array for a quarter long simulation. It uses
-        the quarter number and the hdf5 output file to correct for the time shift
-        of the parent CCD.
-
-        Parameters
-        ----------
-        outputFile : str
-           Full path to the HDF5 outputfile.
-        quarterNo : int
-           Integer specifying which quarter the time array should be created for.
-
-        Return
-        ------
-        time : ndarray 
-            Time array matching the parent quarter defined by the user.
-        """
-
-        # Fetch times
-
-        numExposures = self.hdf5file['InputParameters/ObservingParameters'].attrs['NumExposures']
-        timeStep     = self.hdf5file['InputParameters/ObservingParameters'].attrs['CycleTime']
-        timeShift    = self.hdf5file['InputParameters/CCD'].attrs['TimeShift']
-
-        # Create time array from start and end times
-
-        timeStart = (quarterNo-1) * 90. * 86400. + float(timeShift)
-        timeEnd   = timeStart + timeStep * numExposures
-        time      = np.arange(timeStart, timeEnd, timeStep)
-
-        return time
